added data cleaning step
Discovered the NA issue is caused by using characterize(). Removed characterize() from the .sav upload and removed the .csv upload option.

Updated documentation to reflect the issue with characterize(), in order to use characterize() with SPSS file, we need to make sure all appropriate labels are in the SPSS file before importing.